TIP-1287: move query classes from Doctrine\ORM\Query to Storage\Sql
