+++ src/Akeneo/Pim/Enrichment/Bundle/Storage/Sql/Family/CountProductsWithFamily.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Pim\Enrichment\Bundle\Doctrine\ORM\Query;
+namespace Akeneo\Pim\Enrichment\Bundle\Storage\Sql\Family;
 
 use Akeneo\Pim\Enrichment\Component\Product\Model\ProductInterface;
 use Akeneo\Pim\Enrichment\Component\Product\ProductAndProductModel\Query\CountProductsWithFamilyInterface;
@@ -16,24 +16,16 @@ use Doctrine\ORM\EntityManagerInterface;
  * @copyright 2018 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class CountProductsWithFamily implements CountProductsWithFamilyInterface
+final class CountProductsWithFamily implements CountProductsWithFamilyInterface
 {
     /** @var EntityManagerInterface */
     private $entityManager;
 
-    /**
-     * @param EntityManagerInterface $entityManager
-     */
     public function __construct(EntityManagerInterface $entityManager)
     {
         $this->entityManager = $entityManager;
     }
 
-    /**
-     * @param FamilyInterface $family
-     *
-     * @return int
-     */
     public function count(FamilyInterface $family): int
     {
         $queryBuilder = $this->entityManager->createQueryBuilder();
