+++ src/Akeneo/Pim/Enrichment/Bundle/Storage/Sql/Family/FindAttributesForFamily.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Pim\Enrichment\Bundle\Doctrine\ORM\Query;
+namespace Akeneo\Pim\Enrichment\Bundle\Storage\Sql\Family;
 
 use Akeneo\Pim\Structure\Component\Model\FamilyInterface;
 use Doctrine\DBAL\Driver\Statement;
@@ -20,9 +20,6 @@ class FindAttributesForFamily
     /** @var EntityManagerInterface */
     private $entityManager;
 
-    /**
-     * @param EntityManagerInterface $entityManager
-     */
     public function __construct(EntityManagerInterface $entityManager)
     {
         $this->entityManager = $entityManager;
