+++ src/Akeneo/Pim/Enrichment/Bundle/Storage/Sql/ProductModel/DescendantProductIdsQuery.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Pim\Enrichment\Bundle\Doctrine\ORM\Query;
+namespace Akeneo\Pim\Enrichment\Bundle\Storage\Sql\ProductModel;
 
-use Akeneo\Pim\Enrichment\Component\Product\Query\DescendantProductModelIdsQueryInterface;
+use Akeneo\Pim\Enrichment\Component\Product\Query\DescendantProductIdsQueryInterface;
 use Doctrine\DBAL\Connection;
 
 /**
@@ -12,7 +12,7 @@ use Doctrine\DBAL\Connection;
  * @copyright 2019 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class DescendantProductModelIdsQuery implements DescendantProductModelIdsQueryInterface
+final class DescendantProductIdsQuery implements DescendantProductIdsQueryInterface
 {
     private $connection;
 
@@ -21,16 +21,21 @@ class DescendantProductModelIdsQuery implements DescendantProductModelIdsQueryIn
         $this->connection = $connection;
     }
 
-    public function fetchFromParentProductModelId(int $parentProductModelId): array
+    public function fetchFromProductModelIds(array $productModelIds): array
     {
+        if (empty($productModelIds)) {
+            return [];
+        }
+
         $sql = <<<SQL
-SELECT id FROM pim_catalog_product_model
-WHERE parent_id = :parentId
+SELECT id FROM pim_catalog_product
+WHERE product_model_id IN (:productModelIds)
 SQL;
 
         $resultRows = $this->connection->executeQuery(
             $sql,
-            ['parentId' => $parentProductModelId]
+            ['productModelIds' => $productModelIds],
+            ['productModelIds' => Connection::PARAM_INT_ARRAY]
         )->fetchAll();
 
         return array_map(function ($rowData) {
