+++ tests/back/Pim/Enrichment/Integration/Doctrine/Query/VariantProductRatioIntegration.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 
 namespace AkeneoTest\Pim\Enrichment\Integration\Doctrine\Query;
 
-use Akeneo\Test\Integration\Configuration;
+use keneo\Test\Integration\Configuration;
 use Akeneo\Test\Integration\TestCase;
 
 class VariantProductRatioIntegration extends TestCase
